BE/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = ""
authors = [
    {name = "junho-baek",email = "junho6610@yonsei.ac.kr"}
]
readme = "README.md"

packages = [{include = "be"}]

requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "notion-client (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-dotenv = "^1.0.1"
notion-client = {version = "^2.3.0", python = ">=3.12,<4.0"}
